topic_list: Limit number of unread topics shown at once.
This avoids a stream having potentially near-infinite height when
opened in a stream with a large number of unread topics; the benefit
is that you can easily access the next stream.

We show an unread count next to "more unread topics" to make it hard
to miss.

Needs CSS work to make the count align properly and ideally some
tests.

Fixes #13087.